+++ CHANGELOG.md
@@ -15,4 +15,5 @@ Release notes for the Obfuscator Craft CMS plugin.
 - Added the ability to disable the Reply-To header.
 - Added the ability to specify Reply-To email(s).
 - Added the ability to force plain text only emails.
+- Added the ability to override email body templates.